requirements.txt
# Combined requirements for DarkAI consolidated platform
# Includes dependencies from PersonaForge, BlackWire, and ShadowStack

beautifulsoup4==4.12.2
dnspython==2.4.2
flask==3.0.0
flask-cors==4.0.0
google-search-results==2.4.2  # SerpAPI client
gunicorn==21.2.0
lxml==4.9.3
neo4j==5.15.0
openai>=1.0.0
phonenumbers==8.13.27
playwright==1.40.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-wappalyzer==0.3.1
python-whois==0.8.0
requests==2.31.0
selenium==4.15.2
setuptools>=80.0.0  # Required for python-wappalyzer
